+++ projects/mk-kit/forms/index.ts
@@ -22,6 +22,7 @@ export * from './postal-code-input';
 export * from './currency-input';
 export * from './card-number-input';
 export * from './iban-input';
+export * from './signature-pad';
 export * from './button-toggle';
 export * from './checkbox';
 export * from './radio';